Began transition to new tree structure.

+++ Source/Mesh.h
@@ -41,7 +41,7 @@ typedef struct
 	uint32_t from;
 	uint32_t to;
 	uint32_t next;
-	int64_t other_half;
+	uint32_t other_half;
 } ct_edge_t;
 
 typedef struct
@@ -88,8 +88,8 @@ int ct_mesh_calculate_edges(ct_mesh_t *mesh, char error_message[NM_MAX_ERROR_LEN
 int ct_mesh_check_manifold(ct_mesh_t *mesh, char error_message[NM_MAX_ERROR_LENGTH]);
 uint32_t ct_mesh_triangle_fan_check(ct_mesh_t *mesh, uint32_t vertex, uint32_t vertex_degree);
 uint32_t ct_mesh_get_edge_index(ct_edge_t *edge);
-int64_t ct_mesh_get_next_vertex_edge(ct_mesh_t *mesh, uint32_t vertex, uint32_t edge);
-int64_t ct_mesh_get_previous_vertex_edge(ct_mesh_t *mesh, uint32_t vertex, uint32_t edge);
+uint32_t ct_mesh_get_next_vertex_edge(ct_mesh_t *mesh, uint32_t vertex, uint32_t edge);
+uint32_t ct_mesh_get_previous_vertex_edge(ct_mesh_t *mesh, uint32_t vertex, uint32_t edge);
 
 int ct_mesh_edge_qsort_compare_from(const void *a, const void *b);
 int ct_mesh_edge_qsort_compare_low_high(const void *a, const void *b);
